refactor(SystemStreamRepo):
